Add simple IRC message sender in contrib folder
This commit adds the simple_irc file in the contrib folder. This can be used
for debugging connections to an IRC server and verifying you can send a
message, have proper permissions, authenticaiton, etc. It uses the same ENV VAR
setup as the main program, but removes the requirement of having an inbound
webhook trigger it.