Sizeable refactor
The program now is in separate files. We also have two major services being
provided. The first is the HTTP listener to receive images via an HTTP POST
with an X-API-KEY header. The second looks for images in the storage directory
and puts them in slack.

There are still many steps to go.