feat: API keys can now be issues via a POST on /api
This commit adds an ability to be issued an API key if you provide your slack
Identicier string. This is then used to validate agains the slack team to see
you are a member of said team and then issues you an API key in UUID format.

There are still _some_ security concerns, as getting somebody's slack ID isn't
too difficult, but it's at least an access control mechanism that should limit
the attack vector only members of the slack team.

Right now the only thing implemented is posting to /api with '{"slack_id":
"XXXXXX"}' and if that is validated slack ID, you will be issued an API key.

To test this, you can run `make apirequest` right now.

This will return a simple JSON structure containing your API key.

    {"apikey":"4cb6ed79-e069-4a44-914b-5cebf1085b23","status":"ok"}