feat: Flip _most_ configuration into viper
Previously we had a hybrid of hard-coded, env vars, and viper in play. This
flips most everything to viper. There are still a few variables hanging around
surrounding systemd functionality, but that will get fixed when that is
refactored.

Right now we're setting the following defaults

    DATA_DIR = ./data
    UPLOADS_DIR = ./data/uploads
    DISCARD_DIR = ./data/discard
    PROCESSED_DIR = ./data/processed
    CREDENTIALS_DIR = ./data/credentials
    PORT = 8080

Required for you to set are

    SLACK_TOKEN
    SLACK_CHANNEL