feat: Switch to viper for configuration parsing.
Previously the parsing of configuration was done rather manually through hefty
usage of os.GetEnv and the like. This switches to Viper which is easier to work
with and allows for future expansion into a configuration file or other
options. It's also pretty standard.

At the same time the logger was flipped to logrus.